+++ test/places_test.dart
@@ -1,15 +1,11 @@
-library google_maps_webservice.places.test;
-
 import 'dart:async';
 import 'dart:convert';
-import 'package:http/http.dart';
 import 'package:test/test.dart';
 import 'package:google_maps_webservice/places.dart';
 
-Future<void> launch([Client client]) async {
+Future<void> main() async {
   final apiKey = 'MY_API_KEY';
-  GoogleMapsPlaces places =
-      GoogleMapsPlaces(apiKey: apiKey, httpClient: client);
+  GoogleMapsPlaces places = GoogleMapsPlaces(apiKey: apiKey);
 
   tearDownAll(() {
     places.dispose();
@@ -198,12 +194,15 @@ Future<void> launch([Client client]) async {
                 'https://maps.googleapis.com/maps/api/place/details/json?reference=REF&key=$apiKey'));
       });
 
-      test('with extensions', () {
+      test('with fields', () {
         expect(
-            places.buildDetailsUrl(
-                placeId: 'PLACE_ID', extensions: 'review_summary'),
+            places.buildDetailsUrl(placeId: 'PLACE_ID', fields: [
+              'address_component',
+              'opening_hours',
+              'geometry',
+            ]),
             equals(
-                'https://maps.googleapis.com/maps/api/place/details/json?placeid=PLACE_ID&extensions=review_summary&key=$apiKey'));
+                'https://maps.googleapis.com/maps/api/place/details/json?placeid=PLACE_ID&fields=address_component,opening_hours,geometry&key=$apiKey'));
       });
 
       test('with extensions', () {