+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## 0.0.15
+
+- Handle 'now' string for departure_time
+- Fix alternatives params
+- fix GeocodedWaypoint.partialMatch type (now a boolean)
+
+## 0.0.14
+
+- Support Timezone API, thanks to @aryzhov
+- Support `fields` and `region` for place details
+
+## 0.0.13
+
+- Support Photos API, thanks to @domesticmouse
+
+## 0.0.12
+
+- Add Distance Matrix API, thanks to @1abid
+
 ## 0.0.11
 
 - fix session token